+++ lib/grape.rb
@@ -10,16 +10,16 @@ require 'set'
 require 'active_support'
 require 'active_support/version'
 require 'active_support/isolated_execution_state' if ActiveSupport::VERSION::MAJOR > 6
-require 'active_support/core_ext/hash/indifferent_access'
-require 'active_support/core_ext/object/blank'
+require 'active_support/core_ext/array/conversions'
 require 'active_support/core_ext/array/extract_options'
 require 'active_support/core_ext/array/wrap'
-require 'active_support/core_ext/array/conversions'
-require 'active_support/core_ext/hash/deep_merge'
-require 'active_support/core_ext/hash/reverse_merge'
-require 'active_support/core_ext/hash/except'
-require 'active_support/core_ext/hash/slice'
 require 'active_support/core_ext/hash/conversions'
+require 'active_support/core_ext/hash/deep_merge'
+require 'active_support/core_ext/hash/except'
+require 'active_support/core_ext/hash/indifferent_access'
+require 'active_support/core_ext/hash/reverse_merge'
+require 'active_support/core_ext/hash/slice'
+require 'active_support/core_ext/object/blank'
 require 'active_support/dependencies/autoload'
 require 'active_support/notifications'
 require 'i18n'
@@ -45,6 +45,7 @@ module Grape
     autoload :Env, 'grape/util/env'
     autoload :Json, 'grape/util/json'
     autoload :Xml, 'grape/util/xml'
+    autoload :DryTypes
   end
 
   module Http
@@ -218,6 +219,57 @@ module Grape
       autoload :StreamResponse
     end
   end
+
+  module Validations
+    extend ::ActiveSupport::Autoload
+
+    eager_autoload do
+      autoload :AttributesIterator
+      autoload :MultipleAttributesIterator
+      autoload :SingleAttributeIterator
+      autoload :Types
+      autoload :ParamsScope
+      autoload :ValidatorFactory
+    end
+
+    module Types
+      extend ::ActiveSupport::Autoload
+
+      eager_autoload do
+        autoload :InvalidValue
+        autoload :DryTypeCoercer
+        autoload :ArrayCoercer
+        autoload :SetCoercer
+        autoload :PrimitiveCoercer
+        autoload :CustomTypeCoercer
+        autoload :CustomTypeCollectionCoercer
+        autoload :MultipleTypeCoercer
+        autoload :VariantCollectionCoercer
+      end
+    end
+
+    module Validators
+      extend ::ActiveSupport::Autoload
+
+      eager_autoload do
+        autoload :Base
+        autoload :MultipleParamsBase
+        autoload :AllOrNoneOfValidator
+        autoload :AllowBlankValidator
+        autoload :AsValidator
+        autoload :AtLeastOneOfValidator
+        autoload :CoerceValidator
+        autoload :DefaultValidator
+        autoload :ExactlyOneOfValidator
+        autoload :ExceptValuesValidator
+        autoload :MutualExclusionValidator
+        autoload :PresenceValidator
+        autoload :RegexpValidator
+        autoload :SameAsValidator
+        autoload :ValuesValidator
+      end
+    end
+  end
 end
 
 require 'grape/config'
@@ -226,26 +278,5 @@ require 'grape/content_types'
 require 'grape/util/lazy_value'
 require 'grape/util/lazy_block'
 require 'grape/util/endpoint_configuration'
-
-require 'grape/validations/validators/base'
-require 'grape/validations/attributes_iterator'
-require 'grape/validations/single_attribute_iterator'
-require 'grape/validations/multiple_attributes_iterator'
-require 'grape/validations/validators/allow_blank'
-require 'grape/validations/validators/as'
-require 'grape/validations/validators/at_least_one_of'
-require 'grape/validations/validators/coerce'
-require 'grape/validations/validators/default'
-require 'grape/validations/validators/exactly_one_of'
-require 'grape/validations/validators/mutual_exclusion'
-require 'grape/validations/validators/presence'
-require 'grape/validations/validators/regexp'
-require 'grape/validations/validators/same_as'
-require 'grape/validations/validators/values'
-require 'grape/validations/validators/except_values'
-require 'grape/validations/params_scope'
-require 'grape/validations/validators/all_or_none'
-require 'grape/validations/types'
-require 'grape/validations/validator_factory'
 
 require 'grape/version'