+++ lib/grape/util/json.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
+require 'json'
+
 module Grape
   if Object.const_defined? :MultiJson
     Json = ::MultiJson